+++ examples/evaluate.rs
@@ -1,15 +1,14 @@
-use futures::StreamExt;
-
 use chromiumoxide::browser::{Browser, BrowserConfig};
 use chromiumoxide_cdp::cdp::js_protocol::runtime::{
     CallArgument, CallFunctionOnParams, EvaluateParams,
 };
+use futures::StreamExt;
 
-#[async_std::main]
+#[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let (mut browser, mut handler) = Browser::launch(BrowserConfig::builder().build()?).await?;
 
-    let handle = async_std::task::spawn(async move {
+    let handle = tokio::spawn(async move {
         while let Some(h) = handler.next().await {
             match h {
                 Ok(_) => continue,
@@ -78,6 +77,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     assert_eq!(val, 42);
 
     browser.close().await?;
-    handle.await;
+    handle.await?;
     Ok(())
 }