+++ examples/fetcher.rs
@@ -1,15 +1,14 @@
 use std::path::Path;
 
-use futures::StreamExt;
-
 use chromiumoxide::browser::{Browser, BrowserConfig};
 use chromiumoxide::fetcher::{BrowserFetcher, BrowserFetcherOptions};
+use futures::StreamExt;
 
-#[async_std::main]
+#[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Fetcher browser
     let download_path = Path::new("./download");
-    async_std::fs::create_dir_all(&download_path).await?;
+    tokio::fs::create_dir_all(&download_path).await?;
     let fetcher = BrowserFetcher::new(
         BrowserFetcherOptions::builder()
             .with_path(&download_path)
@@ -25,23 +24,16 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     )
     .await?;
 
-    let handle = async_std::task::spawn(async move {
+    let handle = tokio::spawn(async move {
         loop {
             match handler.next().await {
                 Some(h) => match h {
                     Ok(_) => continue,
-                    Err(e) => {
-                        println!("Err: {}", e);
-                        break;
-                    }
+                    Err(_) => break,
                 },
-                None => {
-                    println!("None");
-                    break;
-                }
+                None => break,
             }
         }
-        println!("Done");
     });
 
     let page = browser.new_page("about:blank").await?;
@@ -51,7 +43,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("it worked!");
 
     browser.close().await?;
-    browser.wait().await?;
-    handle.await;
+    handle.await?;
     Ok(())
 }