+++ examples/wiki.rs
@@ -1,14 +1,14 @@
+use chromiumoxide::browser::{Browser, BrowserConfig};
 use futures::StreamExt;
 
-use chromiumoxide::browser::{Browser, BrowserConfig};
-
-#[async_std::main]
+#[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing_subscriber::fmt::init();
 
-    let (browser, mut handler) = Browser::launch(BrowserConfig::builder().build()?).await?;
+    let (browser, mut handler) =
+        Browser::launch(BrowserConfig::builder().with_head().build()?).await?;
 
-    let handle = async_std::task::spawn(async move {
+    let handle = tokio::spawn(async move {
         loop {
             let _ = handler.next().await.unwrap();
         }
@@ -20,13 +20,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .await?
         .click()
         .await?
-        .type_str("Rust (programming language)")
+        .type_str("Rust programming language")
         .await?
         .press_key("Enter")
         .await?;
 
     let _html = page.wait_for_navigation().await?.content().await?;
 
-    handle.await;
+    handle.await?;
     Ok(())
 }